@@ -1,22 +1,21 @@
 <?php
 /**
- * Zend Framework (http://framework.zend.com/)
+ * Zend Framework (http://framework.zend.com/).
  *
  * @link      http://github.com/zendframework/zf2 for the canonical source repository
+ *
  * @copyright Copyright (c) 2005-2014 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   http://framework.zend.com/license/new-bsd New BSD License
+ *
  * @category  ZendService
- * @package   ZendService_Google
  */
-
 namespace ZendService\Google\Exception;
 
 /**
- * Invalid Argument Exception
+ * Runtime Exception.
  *
  * @category  ZendService
- * @package   ZendService_Google
  */
-class InvalidArgumentException extends \InvalidArgumentException
+class RuntimeException extends \RuntimeException
 {
 }
